Move all nametag methods into their own class
Move all nametag methods from VScene into their own entity class.
